deprecate ssg-manifest
